Support both the legacy and Omnes event systems
We unsubscribe the order mailer subscription from the new event system
when `.disable_builtin_emails` is enabled.

We need to skip testing it for the new event system, although we've
manually checked that it works. The reason is that the best place to
unsubscribe is in an `#after_initialize` hook. As part of the Rails
initialization process, that hook is run only once as part of the dummy
app used in the test suite. The dummy app has `.disable_builtin_emails`
hardcoded to `false` as it's the default when the `solidus_tracking`
generator is run. We'd need to allow that extra flexibility somehow in
the `solidus_dev_support` library.